+++ Sources/Markdown.Portable/LinkStyle.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Xamarin.Forms;
+
+namespace Xam.Forms.Markdown
+{
+    public class LinkStyle : MarkdownStyle
+    {
+        public string OpenLinkSheetTitle { get; set; } = "Open link";
+        public string OpenLinkSheetCancel { get; set; } = "Cancel";
+        public Action<List<LinkData>> CustomTapHandler { get; set; }
+        public List<string> ExternalProtocols = new List<string> { "http://", "https://", "mailto:", "tel:" };
+    }
+
+    public class LinkData
+    {
+        public string Text { get; set; }
+        public string Link { get; set; }
+    }
+
+}