+++ CHANGELOG.md
@@ -3,7 +3,20 @@
 ## Unreleased
 
 ### Added 
+- Support for autolinks (https://spec.commonmark.org/0.29/#autolinks)
+  - Autolinks are rendered the same way as regular links.
+
+## 1.1.0-pre6 - 2021-02-24
+
+### Added 
 - `HorizontalTextAlignment` and `VerticalTextAlignment` for Heading, Paragraph and Code
+- Added `ExternalProtocols` to `LinkStyle` which defines link protocols which can be opened by default tap handler (default are: http://, https://, mailto:, tel:)
+
+### Fixed
+- `CreateSpans` returns `Span[0]` instead of `null` for unknown type, which solves `NullReferenceException`
+
+### Changed
+- `LinkStyle.CustomCallback` renamed to `LinkStyle.CustomTapHandler`
 
 ## 1.1.0-pre5 - 2021-02-18